bot/requirements.txt
python-decouple
aiogram
python-dateutil